requirements.txt
openai>=1.40.0
requests>=2.31.0
python-dateutil>=2.9.0
